Implementation of the Shortest-Job-First algorithm
At this time, the implementation for Shortest-Job-First algorithm (`shortest-job-first.py`) limited to giving the correct answer for a job-set that has a job that starts at time 0. Any set that has a start time other than at 0 shall give incorrect answer (though to be fair, it would be able to give the correct answer for the first job).